new homeWork 04

+++ controllers/contacts.js
@@ -6,17 +6,12 @@ const { HttpError } = require("../helpers");
 
 const listContacts = async (req, res) => {
   const { _id: owner } = req.user;
-  const { page = 1, limit = 10, favorite = false } = req.query;
+  const { page = 1, limit = 10, favorite } = req.query;
   const skip = (page - 1) * limit;
-  const contacts = favorite
-    ? await Contact.find({ owner, favorite: true }, "", {
-        skip,
-        limit,
-      }).populate("owner", "name email")
-    : await Contact.find({ owner }, "", { skip, limit }).populate(
-        "owner",
-        "name email"
-      );
+  const contacts = await Contact.find({ owner, favorite }, "", {
+    skip,
+    limit,
+  }).populate("owner", "name email");
   if (!contacts) {
     throw HttpError(404, `Request failed`);
   }
@@ -25,8 +20,11 @@ const listContacts = async (req, res) => {
 
 const getContactById = async (req, res) => {
   const { id } = req.params;
-  const result = await Contact.findById(id);
-  if (!result) {
+  const { _id: owner } = req.user;
+
+  const result = await Contact.findById(id).populate("owner", "name email");
+
+  if (!result || result.owner._id.toString() !== owner.toString()) {
     throw HttpError(404, `Contact with id: ${id} not found`);
   }
   res.json(result);
@@ -34,7 +32,6 @@ const getContactById = async (req, res) => {
 
 const addContact = async (req, res) => {
   const { _id: owner } = req.user;
-  console.log(owner);
   const result = await Contact.create({ ...req.body, owner });
   if (!result) {
     throw HttpError(404, `Creation failed`);
@@ -44,6 +41,13 @@ const addContact = async (req, res) => {
 
 const removeContact = async (req, res) => {
   const { id } = req.params;
+  const { _id: owner } = req.user;
+
+  const contact = await Contact.findById(id);
+  if (contact.owner.toString() !== owner.toString()) {
+    throw HttpError(404, `Contact with ${id} not found`);
+  }
+
   const result = await Contact.findByIdAndDelete(id);
   if (!result) {
     throw HttpError(404, `Contact with id: ${id} not found`);
@@ -53,6 +57,13 @@ const removeContact = async (req, res) => {
 
 const updateById = async (req, res) => {
   const { id } = req.params;
+  const { _id: owner } = req.user;
+
+  const contact = await Contact.findById(id);
+  if (contact.owner.toString() !== owner.toString()) {
+    throw HttpError(404, `Contact with ${id} not found`);
+  }
+
   const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
   if (!result) {
     throw HttpError(404, `Contact with id: ${id} not found`);
@@ -65,6 +76,13 @@ const updateStatusContact = async (req, res) => {
     return res.status(400).json({ message: "missing fields" });
   }
   const { id } = req.params;
+  const { _id: owner } = req.user;
+
+  const contact = await Contact.findById(id);
+  if (contact.owner.toString() !== owner.toString()) {
+    throw HttpError(404, `Contact with ${id} not found`);
+  }
+
   const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
   if (!result) {
     throw HttpError(404, `Contact with id: ${id} not found`);
